+++ Java/Ema/Examples/src/main/java/com/thomsonreuters/ema/examples/training/iprovider/series200/example280__MarketPrice__Performance/IProvider.java
@@ -125,7 +125,7 @@ class AppClient implements OmmProviderClient
 		
 			System.out.println("total refresh count = " + numberOfRequestItems +
 					"\ttotal time = " + timeSpent + " sec" +
-					"\tupdate rate = " + (float)numberOfRequestItems / timeSpent + " refresh per sec");
+					"\tupdate rate = " + numberOfRequestItems / timeSpent + " refresh per sec");
 		}
 		
 		ReqInfo reqInfo = new ReqInfo();
