+++ Java/Eta/Applications/Examples/src/main/java/com/thomsonreuters/upa/examples/consumer/PostHandler.java
@@ -382,7 +382,7 @@ public class PostHandler
             marketPriceRefresh.state().dataState(DataStates.OK);
             marketPriceRefresh.state().code(StateCodes.NONE);
             marketPriceRefresh.state().text().data("Item Refresh Completed");
-            marketPriceRefresh.serviceId((int)serviceId);
+            marketPriceRefresh.serviceId(serviceId);
             marketPriceRefresh.applyHasServiceId();
             marketPriceRefresh.marketPriceItem(mpItemInfo);
             marketPriceRefresh.applyHasQos();
@@ -613,7 +613,7 @@ public class PostHandler
         postMsg.msgKey().applyHasServiceId();
         postMsg.msgKey().name().data(postItemName.data(), postItemName.position(), postItemName.length());
         postMsg.msgKey().nameType(InstrumentNameTypes.RIC);
-        postMsg.msgKey().serviceId((int)serviceId);
+        postMsg.msgKey().serviceId(serviceId);
 
        	// populate default post user info        
         if (hasInputPrincipalIdentitity)
